% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_xe.R
\name{connect_oracle}
\alias{connect_oracle}
\title{Title}
\usage{
connect_oracle(dbname = "xe", user = "hafvog_user", password = "hafvog")
}
\arguments{
\item{dbname}{The name of the database (dbname), normally "xe" or "XEPDB1"}

\item{user}{XX}

\item{password}{XX}
}
\description{
bla, bla, bla, ...
}
