% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hafvog3.R
\name{import_hafvog3}
\alias{import_hafvog3}
\title{Import data from the Oracle xe-database}
\usage{
import_hafvog3(id = 30, con, year, merge = TRUE, store = FALSE)
}
\arguments{
\item{id}{synaflokkur. Default is set to 30}

\item{con}{Oracle connection}

\item{year}{Current cruise year. If not specificed (default) use current (computer) year.}

\item{merge}{A boolean, if TRUE (default) merge (some) data from {mardata}}

\item{store}{A boolean, if TRUE then the returned returned object is also
saved as hafvog.rds in directory data}
}
\description{
Import data from the Oracle xe-database
}
