% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_smx.R
\name{import_smx}
\alias{import_smx}
\title{Import data from the Oracle xe-database}
\usage{
import_smx(con, schema = c("fiskar", "hafvog"), id = 30, gid = 73)
}
\arguments{
\item{con}{Connection to Oracle, either mar or xe}

\item{schema}{default is "fiskar" and "hafvog". If only interested in reading
from one of them, specify which.}

\item{id}{synaflokkur. Default is set to 30}

\item{gid}{veidarfaeri. Default is set to 73}
}
\description{
Import data from the Oracle xe-database
}
