% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_for_smxapp.R
\name{munge_for_smxapp}
\alias{munge_for_smxapp}
\title{munge_for_smxapp}
\usage{
munge_for_smxapp(res, cruise, rda.file = "smb_dashboard.rda")
}
\arguments{
\item{res}{a list contain st, nu, le and kv and other things. The object is
generated via function import_smx (in the xe-package)}

\item{cruise}{leidangrar in schema hafvog}

\item{rda.file}{name the exported binary file (default smb_dashboard.rda)
which is stored in the directory data2}
}
\description{
Create R-binary bundle for the smxapp
}
