% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xe_fields.R
\name{xe_fields}
\alias{xe_fields}
\title{Get fields/column names of Oracle xe table}
\usage{
xe_fields(con, table)
}
\arguments{
\item{con}{Connection to Oracle xe}

\item{table}{Table name as schema.tablename}
}
\description{
Get fields/column names of Oracle xe table
}
