% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xe_tables.R
\name{xe_tables}
\alias{xe_tables}
\title{Get Oracle xe table names for a scheme}
\usage{
xe_tables(con, schema)
}
\arguments{
\item{con}{Connection to Oracle xe}

\item{schema}{Scheme name}
}
\description{
Get Oracle xe table names for a scheme
}
