% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_smx.R
\name{import_smx}
\alias{import_smx}
\title{Import data from the Oracle xe-database}
\usage{
import_smx(id = 30, gid = 73, year, store = FALSE)
}
\arguments{
\item{id}{synaflokkur. Default is set to 30}

\item{gid}{veidarfaeri. Default is set to 73}

\item{year}{Current cruise year. If not specificed (default) use current (computer) year.}

\item{store}{A boolean, if TRUE then the returned returned object is also
saved as hafvog.rds in directory data}
}
\description{
Import data from the Oracle xe-database
}
