% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_smx.R
\name{import_smx}
\alias{import_smx}
\title{Title}
\usage{
import_smx(con, schema = c("fiskar", "hafvog"), id = 30, gid = 73, cruise,
  debug = 0)
}
\arguments{
\item{con}{XXX}

\item{schema}{XXX}

\item{id}{synaflokkur}

\item{gid}{veidarfaeri}

\item{cruise}{Heiti leiðangurs}

\item{debug}{XXX}
}
\description{
Title
}
